Convert StepController to IdvStepConcern
**Why**: Concerns are better than base class controllers.
