+++ app/controllers/concerns/idv_step_concern.rb
@@ -1,0 +1,32 @@
+module IdvStepConcern
+  extend ActiveSupport::Concern
+
+  include IdvSession
+
+  included do
+    before_action :confirm_two_factor_authenticated
+    before_action :confirm_idv_needed
+    before_action :confirm_idv_session_started
+
+    helper_method :step
+  end
+
+  private
+
+  def increment_step_attempts
+    idv_session.step_attempts[step_name] += 1
+  end
+
+  def step_attempts_exceeded?
+    idv_session.step_attempts[step_name] >= Idv::Attempter.idv_max_attempts
+  end
+
+  def confirm_step_allowed
+    redirect_to_fail_path if step_attempts_exceeded?
+  end
+
+  def redirect_to_fail_path
+    flash[:max_attempts_exceeded] = true
+    redirect_to verify_fail_path
+  end
+end
