+++ app/controllers/verify/finance_other_controller.rb
@@ -1,6 +1,9 @@
 module Verify
-  class FinanceOtherController < StepController
+  class FinanceOtherController < ApplicationController
+    include IdvStepConcern
+
     before_action :confirm_step_needed
+    before_action :confirm_step_allowed
 
     helper_method :idv_finance_form
 
@@ -9,6 +12,10 @@ module Verify
     end
 
     private
+
+    def step_name
+      :financials
+    end
 
     def confirm_step_needed
       redirect_to verify_phone_path if idv_session.financials_confirmation.try(:success?)
