+++ app/controllers/verify/phone_controller.rb
@@ -1,5 +1,7 @@
 module Verify
-  class PhoneController < StepController
+  class PhoneController < ApplicationController
+    include IdvStepConcern
+
     before_action :confirm_step_needed
     before_action :confirm_step_allowed
